Struct grep_searcher::BinaryDetection

pub struct BinaryDetection(/* private fields */);

The behavior of binary detection while searching.

Binary detection is the process of heuristically identifying whether a given chunk of data is binary or not, and then taking an action based on the result of that heuristic. The motivation behind detecting binary data is that binary data often indicates data that is undesirable to search using textual patterns. Of course, there are many cases in which this isn’t true, which is why binary detection is disabled by default.

Unfortunately, binary detection works differently depending on the type of search being executed:

When performing a search using a fixed size buffer, binary detection is applied to the buffer’s contents as it is filled. Binary detection must be applied to the buffer directly because binary files may not contain line terminators, which could result in exorbitant memory usage.

When performing a search using memory maps or by reading data off the heap, then binary detection is only guaranteed to be applied to the parts corresponding to a match. When Quit is enabled, then the first few KB of the data are searched for binary data.

Implementations

impl BinaryDetection

pub fn none() -> BinaryDetection

No binary detection is performed. Data reported by the searcher may contain arbitrary bytes.

This is the default.

pub fn quit(binary_byte: u8) -> BinaryDetection

Binary detection is performed by looking for the given byte.

When searching is performed using a fixed size buffer, then the contents of that buffer are always searched for the presence of this byte. If it is found, then the underlying data is considered binary and the search stops as if it reached EOF.

When searching is performed with the entire contents mapped into memory, then binary detection is more conservative. Namely, only a fixed sized region at the beginning of the contents are detected for binary data. As a compromise, any subsequent matching (or context) lines are also searched for binary data. If binary data is detected at any point, then the search stops as if it reached EOF.

pub fn convert(binary_byte: u8) -> BinaryDetection

Binary detection is performed by looking for the given byte, and replacing it with the line terminator configured on the searcher. (If the searcher is configured to use CRLF as the line terminator, then this byte is replaced by just LF.)

When searching is performed using a fixed size buffer, then the contents of that buffer are always searched for the presence of this byte and replaced with the line terminator. In effect, the caller is guaranteed to never observe this byte while searching.

When searching is performed with the entire contents mapped into memory, then this setting has no effect and is ignored.

pub fn quit_byte(&self) -> Option<u8>

If this binary detection uses the “quit” strategy, then this returns the byte that will cause a search to quit. In any other case, this returns None.

pub fn convert_byte(&self) -> Option<u8>

If this binary detection uses the “convert” strategy, then this returns the byte that will be replaced by the line terminator. In any other case, this returns None.